requirements.txt
alembic==1.0.5
atomicwrites==1.2.1
attrs==18.2.0
boto3==1.9.74
botocore==1.12.74
certifi==2018.11.29
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
docutils==0.14
factory-boy==2.11.1
Faker==1.0.1
Flask==1.0.2
Flask-Env==2.0.0
flask-marshmallow==0.9.0
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.17.0
marshmallow-sqlalchemy==0.15.0
more-itertools==5.0.0
openapi-spec-validator==0.2.4
pluggy==0.8.0
psycopg2-binary==2.7.6.1
py==1.7.0
pytest==4.1.0
pytest-flask==0.14.0
python-dateutil==2.7.5
python-editor==1.0.3
PyYAML==3.13
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
SQLAlchemy==1.2.15
SQLAlchemy-Utils==0.33.10
text-unidecode==1.2
urllib3==1.24.1
Werkzeug==0.14.1
